Move bitmask ops to common
