+++ encord/common/bitmask_operations/__init__.py
@@ -1,5 +1,5 @@
 try:
-    from encord.objects.bitmask_operations.bitmask_operations_numpy import (
+    from encord.common.bitmask_operations.bitmask_operations_numpy import (
         _mask_to_rle,
         _rle_to_mask,
         _rle_to_string,
@@ -9,7 +9,7 @@ try:
         transpose_bytearray,
     )
 except ImportError:
-    from encord.objects.bitmask_operations.bitmask_operations import (
+    from encord.common.bitmask_operations.bitmask_operations import (
         _mask_to_rle,
         _rle_to_mask,
         _rle_to_string,
