+++ tests/objects/test_bitmask_numpy.py
@@ -1,14 +1,14 @@
 import numpy as np
 
-from encord.objects.bitmask import (
-    BitmaskCoordinates,
-)
-from encord.objects.bitmask_operations.bitmask_operations_numpy import (
+from encord.common.bitmask_operations.bitmask_operations_numpy import (
     _mask_to_rle,
     _rle_to_mask,
     _rle_to_string,
     _string_to_rle,
     transpose_bytearray,
+)
+from encord.objects.bitmask import (
+    BitmaskCoordinates,
 )
